Allow float values for sizes (e.g. 0.5em)

+++ stylebot/js/style.js
@@ -203,7 +203,7 @@ stylebot.style = {
         
         var sizeProperties = [ 'font-size', 'line-height', 'letter-spacing', 'letter-height', 'margin', 'margin-top', 'margin-right', 'margin-bottom', 'margin-left', 'padding', 'padding-top', 'padding-right', 'padding-bottom', 'padding-left', 'border-width', 'border-top-width', 'border-right-width', 'border-bottom-width', 'border-left-width' ];
         
-        if ($.inArray(property, sizeProperties) != -1 && parseInt(value))
+        if ($.inArray(property, sizeProperties) != -1 && parseFloat(value))
         {
             if ($.inArray(value, WidgetUI.validSizeUnits) != -1)
                 return false;
